Fixed compiler warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@12010 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/states_screens/cutscene_gui.cpp
@@ -53,7 +53,7 @@ void CutsceneGUI::renderGlobal(float dt)
         core::rect<s32> r(0, UserConfigParams::m_height - GUIEngine::getFontHeight()*2,
                           UserConfigParams::m_width, UserConfigParams::m_height);
         
-        if (GUIEngine::getFont()->getDimension(m_subtitle.c_str()).Width > UserConfigParams::m_width)
+        if (GUIEngine::getFont()->getDimension(m_subtitle.c_str()).Width > (unsigned int)UserConfigParams::m_width)
         {
             GUIEngine::getSmallFont()->draw(m_subtitle, r,
                                             video::SColor(255,255,255,255), true, true, NULL);
